# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ratings:
    post:
      operationId: createRating
      tags: [ratings]
      summary: Rate an artist or venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingInput"
      responses:
        "201":
          description: Rating created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingCreated"
        "429":
          description: Cooldown active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /ratings/{targetType}/{targetId}:
    get:
      operationId: getRatingSummary
      tags: [ratings]
      summary: Get average rating for target
      parameters:
        - in: path
          name: targetType
          required: true
          schema:
            type: string
            enum: [artist, venue]
        - in: path
          name: targetId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rating summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingSummary"
  /venues/me:
    get:
      operationId: getVenueMe
      tags: [venues]
      summary: Get current venue profile
      responses:
        "200":
          description: Venue profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateVenueMe
      tags: [venues]
      summary: Update current venue profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueOnboardingInput"
      responses:
        "200":
          description: Updated venue profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get venue by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Venue:
      type: object
      required: [id, name, address]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        description:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        moods:
          type: array
          items:
            type: string
        imageUrls:
          type: array
          items:
            type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        ownerUsername:
          type: string
          description: Venue account username (for artist messaging)
...
    VenueOnboardingInput:
      type: object
      required: [name, address, lat, lng, moods]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
        address:
          type: string
          minLength: 1
          description: Full address label from geocoding suggestion
        lat:
          type: number
          description: WGS84 latitude from geocoding suggestion
        lng:
          type: number
          description: WGS84 longitude from geocoding suggestion
        description:
          type: string
        size:
          type: string
        moods:
          type: array
          items:
            type: string
        imageUrls:
          type: array
          items:
            type: string
...
    CreateRatingInput:
      type: object
      required: [targetType, targetId, score]
      properties:
        targetType:
          type: string
          enum: [artist, venue]
        targetId:
          type: integer
        score:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    RatingCreated:
      type: object
      required: [id, targetType, targetId, score, createdAt]
      properties:
        id:
          type: integer
        targetType:
          type: string
        targetId:
          type: integer
        score:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RatingSummary:
      type: object
      required: [average, count]
      properties:
        targetType:
          type: string
        targetId:
          type: integer
        average:
          type: number
        count:
          type: integer
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid