# lib/api-zod/src/generated/types/createRatingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRatingInputTargetType } from './createRatingInputTargetType';

export interface CreateRatingInput {
  targetType: CreateRatingInputTargetType;
  targetId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
  comment?: string;
}