# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Venue {
  id: number;
  name: string;
  address: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  size?: string | null;
  moods?: string[];
  imageUrls?: string[];
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** Venue account username (for artist messaging) */
  ownerUsername?: string;
}