# lib/api-zod/src/generated/types/venueOnboardingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VenueOnboardingInput {
  /**
     * @minLength 1
     * @maxLength 60
     */
  name: string;
  /**
     * Full address label from geocoding suggestion
     * @minLength 1
     */
  address: string;
  /** WGS84 latitude from geocoding suggestion */
  lat: number;
  /** WGS84 longitude from geocoding suggestion */
  lng: number;
  description?: string;
  size?: string;
  moods: string[];
  imageUrls?: string[];
}